sequence restructuring
